pyproject.toml
[project]
name = "audio-diffuser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "einops>=0.8.1",
    "einops-exts>=0.0.4",
    "ema-pytorch>=0.7.9",
    "huggingface-hub>=1.3.1",
    "hydra-colorlog==1.2.0",
    "hydra-core==1.3.2",
    "hydra-optuna-sweeper==1.2.0",
    "lightning>=2.0.0",
    "local-attention>=1.8.4",
    "pre-commit>=4.5.1",
    "pyrootutils>=1.0.4",
    "pytest>=9.0.2",
    "pyyaml==6.0",
    "rich>=14.2.0",
    "six==1.16.0",
    "soundfile>=0.13.1",
    "torch>=2.0.0",
    "torchaudio>=2.9.1",
    "torchcodec>=0.9.1",
    "torchmetrics>=0.11.4",
    "torchsde>=0.2.6",
    "torchvision>=0.15.0",
    "wandb>=0.23.1",
]
